requirements.txt
requests==2.31.0
psycopg2-binary
pandas==2.1.1